requirements.txt
atomicwrites==1.2.1
attrs==18.2.0
beautifulsoup4==4.4.1
capstone==4.0.0rc1
catfish==1.4.2
chardet==2.3.0
command-not-found==0.3
defer==1.0.6
enum34==1.1.6
future==0.16.0
html5lib==0.999
isort==4.3.4
language-selector==0.1
lightdm-gtk-greeter-settings==1.2.1
lxml==3.5.0
menulibre==2.1.3
more-itertools==4.3.0
mugshot==0.3.1
onboard==1.2.0
pathlib2==2.3.2
pexpect==4.0.1
Pillow==3.1.2
pluggy==0.7.1
psutil==5.4.7
ptyprocess==0.5
py==1.6.0
py-getch==1.0.1
pycparser==2.18
pycups==1.9.73
pycurl==7.43.0
pyelftools==0.25
Pygments==2.2.0
pygobject==3.20.0
pytest==3.7.4
python-apt==1.1.0b1+ubuntu0.16.4.2
python-debian==0.1.27
python-ptrace==0.9.3
python-systemd==231
pyxdg==0.25
reportlab==3.3.0
requests==2.9.1
ROPGadget==5.4
sessioninstaller==0.0.0
six==1.11.0
termcolor==1.1.0
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
unicorn==1.0.1
urllib3==1.13.1
xkit==0.0.0
